extract component getters into helpers
