+++ src/components/victory-chart.jsx
@@ -18,6 +18,7 @@ import {
 } from "victory-util";
 import { VictoryAxis } from "victory-axis";
 import { VictoryLine } from "victory-line";
+import * as Helpers from "../helper-methods";
 
 const defaultAxes = {
   independent: <VictoryAxis animate={{velocity: 0.02}}/>,
@@ -132,107 +133,6 @@ export default class VictoryChart extends React.Component {
       },
       styleProps
     )};
-  }
-
-  getAxisType(child) {
-    if (!child.type || child.type.role !== "axis") {
-      return undefined;
-    }
-    return child.props.dependentAxis ? "dependent" : "independent";
-  }
-
-  getChildComponents(props) {
-    // set up a counter for component types
-    const count = () => {
-      const counts = {};
-      return {
-        add: (child) => {
-          const type = child.type && child.type.role;
-          const axis = this.getAxisType(child);
-          if (!counts[type]) {
-            counts[type] = axis ? {independent: 0, dependent: 0} : 0;
-          }
-          if (axis) {
-            counts[type][axis] = counts[type][axis] += 1;
-          } else {
-            counts[type] = counts[type] += 1;
-          }
-        },
-        limitReached: (child) => {
-          const type = child.type && child.type.role;
-          const axis = this.getAxisType(child);
-          if (!counts[type]) {
-            return false;
-          } else if (axis) {
-            return counts[type][axis] > 1;
-          } else if (type === "bar") {
-            // TODO: should we remove the limit on grouped data types?
-            return counts[type] > 1;
-          }
-          return false;
-        },
-        total: (type, axis) => {
-          const totalCount = (axis && counts[type]) ?
-            counts[type][axis] : counts[type];
-          return totalCount || 0;
-        }
-      };
-    }();
-
-    if (!props.children) {
-      return [defaultData, defaultAxes.independent, defaultAxes.dependent];
-    }
-    const childComponents = [];
-    // loop through children, and add each child to the childComponents array
-    // unless the limit for that child type has already been reached.
-    React.Children.forEach(props.children, (child) => {
-      if (!child || !child.type) { return; }
-      const type = child.type && child.type.role;
-      if (count.limitReached(child)) {
-        const msg = type === "axis" ?
-          `Only one VictoryAxis component of each axis type is allowed when using the ` +
-          `VictoryChart wrapper. Only the first axis will be used. Please compose ` +
-          `multi-axis charts manually` :
-          `Only one " + type + "component is allowed per chart. If you are trying ` +
-          `to plot several datasets, please pass an array of data arrays directly ` +
-          `into ${type}.`;
-        Log.warn(msg);
-      }
-      childComponents.push(child);
-      count.add(child);
-    });
-
-    // Add default axis components if necessary
-    // TODO: should we add both axes by default?
-    if (count.total("axis", "independent") < 1) {
-      childComponents.push(defaultAxes.independent);
-    }
-    if (count.total("axis", "dependent") < 1) {
-      childComponents.push(defaultAxes.dependent);
-    }
-
-    // Add defaut data if no data is provided
-    const dataComponents = childComponents.filter((child) => {
-      const type = child.type && child.type.role;
-      return type !== "axis";
-    });
-
-    if (dataComponents.length === 0) { childComponents.push(defaultData); }
-    return childComponents;
-  }
-
-  getDataComponents(childComponents) {
-    return childComponents.filter((child) => {
-      const type = child.type && child.type.role;
-      return type !== "bar" && type !== "axis";
-    });
-  }
-
-  getGroupedDataComponents(childComponents) {
-    return childComponents.filter((child) => {
-      const type = child.type && child.type.role;
-      return type === "bar";
-    });
   }
 
   createStringMap(childComponents, axis) {
@@ -508,13 +408,13 @@ export default class VictoryChart extends React.Component {
   }
 
   getAllComponents(props) {
-    const all = this.getChildComponents(props);
+    const all = Helpers.getChildComponents(props, defaultData, defaultAxes);
     const axis = {
       x: this.getAxisComponent(all, "x"),
       y: this.getAxisComponent(all, "y")
     };
-    const data = this.getDataComponents(all);
-    const groupedData = this.getGroupedDataComponents(all);
+    const data = Helpers.getDataComponents(all, "data");
+    const groupedData = Helpers.getDataComponents(all, "grouped");
     return { all, axis, data, groupedData };
   }
